+++ lib/src/data_source/deltalake.rs
@@ -5,15 +5,16 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use datafusion::arrow::array::RecordBatch;
+use datafusion::arrow::{array::RecordBatch, compute};
 use delta_kernel::{
     engine::{
         arrow_data::ArrowEngineData,
         default::{executor::tokio::TokioBackgroundExecutor, DefaultEngine},
     },
     scan::ScanBuilder,
-    Engine, Table,
+    DeltaResult, Table,
 };
+use itertools::Itertools;
 use object_store::DynObjectStore;
 
 use crate::data_source::location::{
@@ -47,17 +48,17 @@ pub fn to_record_batch(
     let table = Table::try_from_uri(location)
         .map_err(|e| ResponseError::request_validation(e.to_string()))?;
 
-    let engine: Box<dyn Engine> = if scheme == SupportedScheme::File {
+    let engine = Arc::new(if scheme == SupportedScheme::File {
         let mut options: HashMap<&str, String> = HashMap::new();
         options.insert("skip_signature", "true".to_string()); // no cloud credentials are needed
 
-        Box::new(DefaultEngine::try_new(
+        DefaultEngine::try_new(
             table.location(),
             options,
             Arc::new(TokioBackgroundExecutor::new()),
-        )?)
+        )?
     } else {
-        Box::new(DefaultEngine::new(
+        DefaultEngine::new(
             build_store(
                 &scheme,
                 parts.authority.as_ref().map_or("", |auth| auth.as_str()),
@@ -66,23 +67,30 @@ pub fn to_record_batch(
                 parts.path_and_query.as_ref().map_or("/", |pq| pq.path()),
             ),
             Arc::new(TokioBackgroundExecutor::new()),
-        ))
-    };
+        )
+    });
 
     let snapshot = table.snapshot(engine.as_ref(), options.version)?;
     let scan = ScanBuilder::new(snapshot).build()?;
 
-    let mut batches = vec![];
-
-    for res in scan.execute(engine.as_ref())? {
-        let record_batch: RecordBatch = res
-            .raw_data?
-            .into_any()
-            .downcast::<ArrowEngineData>()
-            .map_err(|_| delta_kernel::Error::EngineDataType("ArrowEngineData".to_string()))?
-            .into();
-        batches.push(record_batch);
-    }
+    let batches: Vec<RecordBatch> = scan
+        .execute(engine)?
+        .map(|scan_result| -> DeltaResult<_> {
+            let scan_result = scan_result?;
+            let mask = scan_result.full_mask();
+            let data = scan_result.raw_data?;
+            let record_batch: RecordBatch = data
+                .into_any()
+                .downcast::<ArrowEngineData>()
+                .map_err(|_| delta_kernel::Error::EngineDataType("ArrowEngineData".to_string()))?
+                .into();
+            if let Some(mask) = mask {
+                Ok(compute::filter_record_batch(&record_batch, &mask.into())?)
+            } else {
+                Ok(record_batch)
+            }
+        })
+        .try_collect()?;
 
     Ok(batches)
 }